feat: edit directory structure for viewmodel(MVVM) & add JoinViewModel
