+++ src/mobile/app/src/main/java/com/smilegate/Easel/presentation/view/join/NeedPasswordFragment.kt
@@ -1,4 +1,4 @@
-package com.smilegate.Easel.ui.login
+package com.smilegate.Easel.presentation.view.join
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,13 +7,10 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.NavController
 import androidx.navigation.fragment.findNavController
-import com.smilegate.Easel.R
-import com.smilegate.Easel.databinding.FragmentLoginBinding
-import com.smilegate.Easel.databinding.FragmentStartBinding
+import com.smilegate.Easel.databinding.FragmentNeedPasswordBinding
 
-
-class LoginFragment : Fragment() {
-    private lateinit var binding: FragmentLoginBinding
+class NeedPasswordFragment : Fragment() {
+    private lateinit var binding: FragmentNeedPasswordBinding
 
     private lateinit var navController: NavController
 
@@ -21,17 +18,10 @@ class LoginFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentLoginBinding.inflate(inflater, container, false)
+        binding = FragmentNeedPasswordBinding.inflate(inflater, container, false)
 
         navController = findNavController()
 
-        binding.loginFragmentNextBtn.setOnClickListener {
-            navController.navigate(R.id.action_loginFragment_to_passwordFragment)
-        }
-
-        binding.loginFragmentLostPw.setOnClickListener {
-            navController.navigate(R.id.action_loginFragment_to_findAccountFragment)
-        }
 
         return binding.root
     }
