+++ src/mobile/app/src/main/java/com/smilegate/Easel/presentation/view/login/LoginFragment.kt
@@ -1,4 +1,4 @@
-package com.smilegate.Easel.ui.join
+package com.smilegate.Easel.presentation.view.login
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,11 +8,11 @@ import android.view.ViewGroup
 import androidx.navigation.NavController
 import androidx.navigation.fragment.findNavController
 import com.smilegate.Easel.R
-import com.smilegate.Easel.databinding.FragmentNeedPasswordBinding
-import com.smilegate.Easel.databinding.FragmentStartBinding
+import com.smilegate.Easel.databinding.FragmentLoginBinding
 
-class NeedPasswordFragment : Fragment() {
-    private lateinit var binding: FragmentNeedPasswordBinding
+
+class LoginFragment : Fragment() {
+    private lateinit var binding: FragmentLoginBinding
 
     private lateinit var navController: NavController
 
@@ -20,10 +20,17 @@ class NeedPasswordFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentNeedPasswordBinding.inflate(inflater, container, false)
+        binding = FragmentLoginBinding.inflate(inflater, container, false)
 
         navController = findNavController()
 
+        binding.loginFragmentNextBtn.setOnClickListener {
+            navController.navigate(R.id.action_loginFragment_to_passwordFragment)
+        }
+
+        binding.loginFragmentLostPw.setOnClickListener {
+            navController.navigate(R.id.action_loginFragment_to_findAccountFragment)
+        }
 
         return binding.root
     }
