+++ src/mobile/app/src/main/java/com/smilegate/Easel/presentation/view/login/PasswordFragment.kt
@@ -1,20 +1,14 @@
-package com.smilegate.Easel.ui.login
+package com.smilegate.Easel.presentation.view.login
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import androidx.navigation.NavController
 import androidx.navigation.fragment.findNavController
 import com.smilegate.Easel.R
-import com.smilegate.Easel.databinding.FragmentLoginBinding
 import com.smilegate.Easel.databinding.FragmentPasswordBinding
-
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
