+++ src/mobile/app/src/main/res/layout/fragment_create_account.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:backgroundTint="@color/white"
-    tools:context=".ui.join.CreateAccountFragment">
+    tools:context=".presentation.view.join.CreateAccountFragment">
 
     <TextView
         android:id="@+id/textView"
