+++ src/mobile/app/src/main/res/layout/fragment_find_account.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:backgroundTint="@color/white"
-    tools:context=".ui.login.FindAccountFragment">
+    tools:context=".presentation.view.login.FindAccountFragment">
 
     <TextView
         android:id="@+id/textView"
