+++ src/mobile/app/src/main/res/layout/fragment_need_password.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:backgroundTint="@color/white"
-    tools:context=".ui.join.NeedPasswordFragment">
+    tools:context=".presentation.view.join.NeedPasswordFragment">
 
     <TextView
         android:id="@+id/textView"
