+++ src/mobile/app/src/main/res/layout/fragment_password.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:backgroundTint="@color/white"
-    tools:context=".ui.login.PasswordFragment">
+    tools:context=".presentation.view.login.PasswordFragment">
 
     <TextView
         android:id="@+id/textView"
