+++ src/mobile/app/src/main/res/layout/fragment_send_code.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:backgroundTint="@color/white"
-    tools:context=".ui.join.SendCodeFragment">
+    tools:context=".presentation.view.join.SendCodeFragment">
 
     <TextView
         android:id="@+id/textView"
@@ -29,7 +29,6 @@
         android:layout_marginTop="54dp"
         android:layout_marginEnd="24dp"
         android:background="@null"
-        android:inputType="number"
         android:textColor="@color/BlueGrey_800"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
