+++ src/mobile/app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/startFragment"
-        android:name="com.smilegate.Easel.ui.join.StartFragment"
+        android:name="com.smilegate.Easel.presentation.view.join.StartFragment"
         android:label="fragment_start"
         tools:layout="@layout/fragment_start">
 
@@ -23,7 +23,7 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.smilegate.Easel.ui.login.LoginFragment"
+        android:name="com.smilegate.Easel.presentation.view.login.LoginFragment"
         android:label="fragment_login"
         tools:layout="@layout/fragment_login">
 
@@ -39,7 +39,7 @@
 
     <fragment
         android:id="@+id/passwordFragment"
-        android:name="com.smilegate.Easel.ui.login.PasswordFragment"
+        android:name="com.smilegate.Easel.presentation.view.login.PasswordFragment"
         android:label="fragment_password"
         tools:layout="@layout/fragment_password">
 
@@ -51,7 +51,7 @@
 
     <fragment
         android:id="@+id/findAccountFragment"
-        android:name="com.smilegate.Easel.ui.login.FindAccountFragment"
+        android:name="com.smilegate.Easel.presentation.view.login.FindAccountFragment"
         android:label="fragment_fine_account"
         tools:layout="@layout/fragment_find_account">
 
@@ -59,7 +59,7 @@
 
     <fragment
         android:id="@+id/createAccountFragment"
-        android:name="com.smilegate.Easel.ui.join.CreateAccountFragment"
+        android:name="com.smilegate.Easel.presentation.view.join.CreateAccountFragment"
         android:label="fragment_create_account"
         tools:layout="@layout/fragment_create_account">
 
@@ -71,7 +71,7 @@
 
     <fragment
         android:id="@+id/sendCodeFragment"
-        android:name="com.smilegate.Easel.ui.join.SendCodeFragment"
+        android:name="com.smilegate.Easel.presentation.view.join.SendCodeFragment"
         android:label="fragment_send_code"
         tools:layout="@layout/fragment_send_code">
 
@@ -83,7 +83,7 @@
 
     <fragment
         android:id="@+id/needPasswordFragment"
-        android:name="com.smilegate.Easel.ui.join.NeedPasswordFragment"
+        android:name="com.smilegate.Easel.presentation.view.join.NeedPasswordFragment"
         android:label="fragment_need_password"
         tools:layout="@layout/fragment_need_password">
